+++ src/main/java/troublog/backend/domain/trouble/dto/request/PostUpdateReqDto.java
@@ -86,7 +86,7 @@ public record PostUpdateReqDto(
 		example = "1"
 	)
 	@NotNull(message = "Project ID는 null 일 수 없습니다.")
-	Integer projectId,
+	Long projectId,
 
 	@Schema(
 		description = "썸네일 이미지 URL",
